requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk == 3.9.1
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box ==6.0.2
ensure == 1.0.2


-e .